# [MRG] Fix OverflowError on DictVectorizer