@TomDLT 2019-11-02
Thanks for the pull-request !
Could you add a small test in `sklearn/feature_extraction/tests/test_dict_vectorizer.py` ?