@rth 2019-11-02
> Could you add a small test in sklearn/feature_extraction/tests/test_dict_vectorizer.py ?

The problem is that this only happens for sparse arrays that have more than 2**31 elements which corresponds to 10s GB RAM. I think it's OK to merge it without unit test given that's idential to what we did earlier for `CountVectorizer` in https://github.com/scikit-learn/scikit-learn/pull/9147