@rth 2019-11-02
@norvan Could you please run a small performance benchmark before and after this PR to make sure performance is not degraded? It shouldn't be, but better to double check.

For instance running evaluation code in jupyter with `%timeit DictVectorizer` should be enough. Thanks!