@norvan 2019-11-02
Ran the following code on both stable sklearn and after my changes.
Duration was:
* `2.79 s ± 16.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)` before the change
* `2.83 s ± 25.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)` after the change

Conclusion is that the change does not impair performance in a statistically meaningful way considering the tolerance.

```python
from sklearn.feature_extraction import DictVectorizer
from sklearn.feature_selection import GenericUnivariateSelect, chi2
from numpy.random import random

n_samples = 1000000
do_print = True

data = [{"alpha": i, "beta": int(random()*100)} for i in range(n_samples)]
labels = list(range(n_samples))

vectorizer = DictVectorizer(sparse=True, dtype=np.uint16).fit(data)

%timeit vectorized_data = vectorizer.transform(data)
```